changed voxel density to float and MC CPU working

+++ Assets/VoxelMaster/Core/Voxel.cs
@@ -1,5 +1,5 @@
 public struct Voxel
 {
-    public byte density;
+    public float density;
     public byte materialIndex;
 }
